wip refactor geoportalmap
