+++ apps/geoportal/src/app/components/GeoportalMap/GeoportalMap.tsx
@@ -1,106 +1,32 @@
-import L from "leaflet";
-import proj4 from "proj4";
-import {
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useContext, useEffect, useMemo, useRef } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { Button, Tooltip } from "antd";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faArrowRightFromBracket,
-  faKey,
-} from "@fortawesome/free-solid-svg-icons";
-
 import { TopicMapContext } from "react-cismap/contexts/TopicMapContextProvider";
-import { UIDispatchContext } from "react-cismap/contexts/UIContextProvider";
-import TopicMapComponent from "react-cismap/topicmaps/TopicMapComponent";
-import GenericModalApplicationMenu from "react-cismap/topicmaps/menu/ModalApplicationMenu";
+
+import { useGeoportalOverlays } from "./utils/useOverlayHelpers";
+import type { LeafletConfig } from "@carma/types";
 
 import {
-  SelectionItem,
-  TopicMapSelectionContent,
-  useGazData,
-  useSelectionCesium,
-  useSelectionTopicMap,
-  useCesiumModels,
-} from "@carma-appframeworks/portals";
-import {
-  geoElements,
-  getCollabedHelpComponentConfig,
-  tooltipText,
-} from "@carma-collab/wuppertal/geoportal";
-import { getCollabedHelpComponentConfig as getCollabedHelpElementsConfig } from "@carma-collab/wuppertal/helper-overlay";
-
-import { ENDPOINT, isAreaType } from "@carma-commons/resources";
-import type { FeatureInfo } from "@carma/types";
-
-import {
-  useOverlayHelper,
-  useOverlayTourContext,
-} from "@carma-commons/ui/helper-overlay";
-import { getApplicationVersion } from "@carma-commons/utils";
-
-import {
-  CustomViewer,
-  selectShowPrimaryTileset,
   selectViewerIsMode2d,
-  selectViewerModels,
   setCurrentSceneStyle,
   useCesiumContext,
-  useCesiumInitialCameraFromSearchParams,
 } from "@carma-mapping/engines/cesium";
-import { EmptySearchComponent } from "@carma-mapping/fuzzy-search";
-import { useAuth } from "@carma-providers/auth";
-import { useFeatureFlags } from "@carma-providers/feature-flag";
 
-import FeatureInfoBox from "../feature-info/FeatureInfoBox.tsx";
-import InfoBoxMeasurement from "../map-measure/InfoBoxMeasurement.jsx";
-import PrintPreview from "../map-print/PrintPreview.tsx";
-
-import versionData from "../../../version.json";
-
-import { proj4crs3857def, proj4crs4326def } from "../../helper/gisHelper.js";
-import { getBackgroundLayers } from "../../helper/layer.tsx";
-import { addCssToOverlayHelperItem } from "../../helper/overlayHelper.ts";
-
-import { useLeafletZoomControls } from "../../hooks/leaflet/useLeafletZoomControls.ts";
 import { useDispatchSachdatenInfoText } from "../../hooks/useDispatchSachdatenInfoText.ts";
-import { useFeatureInfoModeCursorStyle } from "../../hooks/useFeatureInfoModeCursorStyle.ts";
-import { useObliqueInitializer } from "../../oblique/hooks/useObliqueInitializer.ts";
 
-import { onClickTopicMap } from "./topicmap.utils.ts";
-import { useCreateCismapLayers } from "./hooks/useCreateCismapLayer.ts";
-
-import store from "../../store/index.ts";
-import {
-  getLoading,
-  getSelectedFeature,
-  setFeatures,
-  setPreferredLayerId,
-  setSecondaryInfoBoxElements,
-  setSelectedFeature,
-} from "../../store/slices/features.ts";
 import {
   getBackgroundLayer,
-  getLayers,
   getLayersIdle,
-  getShowHamburgerMenu,
   setLayersIdle,
 } from "../../store/slices/mapping.ts";
-import { getUIMode, UIMode } from "../../store/slices/ui.ts";
-
-import LoginForm from "../LoginForm.tsx";
-import { useModelSelectionDispatcher } from "../../hooks/useModelSelectionDispatcher.ts";
+// no UI mode handling here; wrapper handles 2D-specific UI flows
 
 import { useLocationChangeHandlers } from "./hooks/useLocationChangeHandlers.ts";
+import { TopicMapComponentWrapper } from "./components/TopicMapComponentWrapper";
+import CesiumMapComponentWrapper from "./components/CesiumMapComponentWrapper.tsx";
 
 import { CESIUM_CONFIG, LEAFLET_CONFIG } from "../../config/app.config";
+import type { CesiumConfig } from "@carma-mapping/engines/cesium";
 
 import "cesium/Build/Cesium/Widgets/widgets.css";
 import "../leaflet.css";
@@ -109,59 +35,37 @@ interface MapProps {
   height: number;
   width: number;
   allow3d?: boolean;
+  options?: {
+    topicmap?: Partial<LeafletConfig>;
+    cesium?: Partial<CesiumConfig>;
+  };
 }
 
-export const GeoportalMap = ({ height, width, allow3d }: MapProps) => {
+export const GeoportalMap = ({
+  height: _height,
+  width: _width,
+  allow3d,
+  options,
+}: MapProps) => {
   const dispatch = useDispatch();
+
+  const cesiumOptions = options?.cesium ?? CESIUM_CONFIG;
+  const leafletOptions = options?.topicmap ?? LEAFLET_CONFIG;
 
   // Contexts
   const ctx = useCesiumContext();
-  const { withTerrainProvider, withSurfaceProvider } = ctx;
+  // using ctx.isValidViewer() below
 
   const rerenderCountRef = useRef(0);
   const lastRenderTimeStampRef = useRef(Date.now());
   const lastRenderIntervalRef = useRef(0);
-  const container3dMapRef = useRef<HTMLDivElement>(null);
+  // 3D container is handled inside CesiumMapComponentWrapper
 
   // State and Selectors
   const backgroundLayer = useSelector(getBackgroundLayer);
   const isMode2d = useSelector(selectViewerIsMode2d) || !allow3d;
-  const models = useSelector(selectViewerModels);
-  const markerAsset = models[CESIUM_CONFIG.markerKey]; //
-  const markerAnchorHeight = CESIUM_CONFIG.markerAnchorHeight ?? 10;
-  const layers = useSelector(getLayers);
-  const uiMode = useSelector(getUIMode);
-  const isModeMeasurement = uiMode === UIMode.MEASUREMENT;
-  const isModeFeatureInfo = uiMode === UIMode.FEATURE_INFO;
-  const showHamburgerMenu = useSelector(getShowHamburgerMenu);
-  const selectedFeature = useSelector(getSelectedFeature);
-  const loadingFeatureInfo = useSelector(getLoading);
-  const { jwt, setJWT } = useAuth();
 
-  const { getLeafletZoom } = useLeafletZoomControls();
-  const showPrimaryTileset = useSelector(selectShowPrimaryTileset);
-
-  const infoBoxOverlay = addCssToOverlayHelperItem(
-    getCollabedHelpElementsConfig("INFOBOX", geoElements),
-    "350px",
-    "137px"
-  );
-
-  const layerButtonsOverlay = addCssToOverlayHelperItem(
-    getCollabedHelpElementsConfig("LAYERBUTTONS", geoElements),
-    "146px",
-    "21px"
-  );
-
-  const mapInteractionOverlay = addCssToOverlayHelperItem(
-    getCollabedHelpElementsConfig("CENTER", geoElements),
-    "15px",
-    "15px"
-  );
-
-  useOverlayHelper(infoBoxOverlay);
-  useOverlayHelper(layerButtonsOverlay);
-  useOverlayHelper(mapInteractionOverlay);
+  useGeoportalOverlays();
 
   const { routedMapRef: topicMap } =
     useContext<typeof TopicMapContext>(TopicMapContext);
@@ -192,140 +96,9 @@ export const GeoportalMap = ({ height, width, allow3d }: MapProps) => {
   const { topicMapLocationChangedHandler, cesiumLocationChangedHandler } =
     useLocationChangeHandlers(isMode2d, changeHandlerOptions);
 
-  const { setAppMenuVisible } =
-    useContext<typeof UIDispatchContext>(UIDispatchContext);
-  const { setSecondaryWithKey, showOverlayHandler } = useOverlayTourContext();
-
-  const [isLoginFormVisible, setIsLoginFormVisible] = useState(false);
-  const [marker, setMarker] = useState(undefined);
-  const [markerAccent, setMarkerAccent] = useState(undefined);
-  const [pos, setPos] = useState<[number, number] | null>(null);
-  // TODO: move all these to a custom hook and collect all calls to updateFeatureInfo there
-  const [shouldUpdateFeatureInfo, setShouldUpdateFeatureInfo] =
-    useState<boolean>(false);
-  const version = getApplicationVersion(versionData);
-
-  // custom hooks
-  const flags = useFeatureFlags();
-  const { isDebugMode } = flags;
-
-  const cesiumInitialCameraView = useCesiumInitialCameraFromSearchParams();
-  const { isObliqueMode } = useObliqueInitializer(isDebugMode);
-
   useDispatchSachdatenInfoText();
-  const modelSelectionDispatcher = useModelSelectionDispatcher();
-
-  const modelConfig = useMemo(
-    () => ({
-      models: CESIUM_CONFIG.models || [],
-      enabled: flags.featureFlagBugaBridge && !isMode2d,
-      selection: {
-        enabled: flags.featureFlagBugaBridge && !isMode2d,
-        deselectOnEmptyClick: true,
-        onSelect: (feature) =>
-          modelSelectionDispatcher(feature as FeatureInfo | null),
-      },
-    }),
-    [flags.featureFlagBugaBridge, isMode2d, modelSelectionDispatcher]
-  );
-
-  useCesiumModels(modelConfig);
-
-  const { gazData } = useGazData();
-
-  useFeatureInfoModeCursorStyle();
-
-  const onComplete = useCallback(
-    (selection: SelectionItem) => {
-      if (layers.filter((l) => l.layerType === "vector").length === 0) return;
-      if (
-        (uiMode === UIMode.DEFAULT || uiMode === UIMode.FEATURE_INFO) &&
-        !isAreaType(selection.type as ENDPOINT) &&
-        isMode2d
-      ) {
-        const selectedPos = proj4(proj4crs3857def, proj4crs4326def, [
-          selection.x,
-          selection.y,
-        ]);
-        if (layersIdle) {
-          const map = getTopicMap();
-          const updatedPos = { lat: selectedPos[1], lng: selectedPos[0] };
-          const latlngPoint = L.latLng(updatedPos);
-
-          if (map) {
-            map.fireEvent("click", {
-              latlng: latlngPoint,
-              layerPoint: map.latLngToLayerPoint(latlngPoint),
-              containerPoint: map.latLngToContainerPoint(latlngPoint),
-            });
-          }
-        } else {
-          setTimeout(() => {
-            onComplete(selection);
-          }, 20);
-        }
-      }
-    },
-    [layers, uiMode, isMode2d, layersIdle, getTopicMap]
-  );
-
-  const updateFeatureInfoLeaflet = useCallback(() => {
-    setShouldUpdateFeatureInfo(false);
-    if (!pos) return;
-
-    setTimeout(() => {
-      const latlngPoint = L.latLng(pos);
-      const map = getTopicMap();
-      if (map) {
-        map.fireEvent("click", {
-          latlng: latlngPoint,
-          layerPoint: map.latLngToLayerPoint(latlngPoint),
-          containerPoint: map.latLngToContainerPoint(latlngPoint),
-        });
-      }
-    }, 150);
-  }, [pos, getTopicMap]);
-
-  useSelectionTopicMap({ onComplete });
-
-  const markerConfig = useMemo(
-    () => ({
-      markerAsset,
-      markerAnchorHeight,
-      isPrimaryStyle: showPrimaryTileset,
-      withTerrainProvider,
-      withSurfaceProvider,
-    }),
-    [
-      markerAsset,
-      markerAnchorHeight,
-      showPrimaryTileset,
-      withTerrainProvider,
-      withSurfaceProvider,
-    ]
-  );
-
-  useSelectionCesium(!isMode2d, markerConfig, isObliqueMode);
 
   useEffect(() => {
-    let layerType = "";
-
-    if (layers.length === 0) {
-      dispatch(setSecondaryInfoBoxElements([]));
-      dispatch(setFeatures([]));
-      dispatch(setSelectedFeature(null));
-    }
-
-    layers.forEach((layer, i) => {
-      if (i === 0) {
-        layerType = layer.layerType;
-      }
-    });
-  }, [layers]);
-
-  useEffect(() => {
-    // TODO wrap this with 3d component in own component?
-    // INTIALIZE Cesium Tileset style from Geoportal/TopicMap background later style
     if (ctx.isValidViewer() && backgroundLayer) {
       if (backgroundLayer.id === "luftbild") {
         dispatch(setCurrentSceneStyle("primary"));
@@ -336,198 +109,17 @@ export const GeoportalMap = ({ height, width, allow3d }: MapProps) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [backgroundLayer]);
 
-  useEffect(() => {
-    const map = routedMap?.leafletMap?.leafletElement;
-    if (uiMode !== UIMode.FEATURE_INFO && marker !== undefined && map) {
-      map.removeLayer(marker);
-      map.removeLayer(markerAccent);
-      dispatch(setSelectedFeature(null));
-      dispatch(setSecondaryInfoBoxElements([]));
-      dispatch(setFeatures([]));
-      setPos(null);
-      dispatch(setPreferredLayerId(""));
-    }
-  }, [uiMode]);
+  // UI mode-based marker cleanup handled in the wrapper
 
-  useEffect(() => {
-    if (isModeFeatureInfo && pos) updateFeatureInfoLeaflet();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [layers]);
+  // Feature info refresh on layer changes handled in the wrapper
 
-  useEffect(() => {
-    const handleZoomEnd = () => {
-      setShouldUpdateFeatureInfo(true);
-    };
+  // Zoom-based feature info refresh handled in the wrapper
 
-    const map = getTopicMap();
+  // InfoBox rendering is handled in the wrapper
 
-    if (!map) return;
+  // Overlay-trigger shortcut moved into wrapper
 
-    map.on("zoomend", handleZoomEnd);
-
-    // Clean up the event listener when the component unmounts
-    return () => {
-      const map = getTopicMap();
-      if (!map) return;
-      map.off("zoomend", handleZoomEnd);
-    };
-  }, [getTopicMap]);
-
-  const renderInfoBox = useCallback(() => {
-    if (isMode2d) {
-      if (isModeMeasurement) {
-        return <InfoBoxMeasurement key={uiMode} />;
-      }
-      if (selectedFeature || loadingFeatureInfo) {
-        return <FeatureInfoBox pos={pos} />;
-      }
-    } else if (flags.featureFlagBugaBridge && selectedFeature) {
-      // TODO unify with point queries for position information?
-      return <FeatureInfoBox />;
-    }
-
-    return <div></div>;
-  }, [
-    isMode2d,
-    isModeMeasurement,
-    selectedFeature,
-    loadingFeatureInfo,
-    pos,
-    flags.featureFlagBugaBridge,
-    uiMode,
-  ]);
-
-  const showOverlayFromOutside = useCallback(
-    (key: string) => {
-      setAppMenuVisible(false);
-      setSecondaryWithKey(key);
-      showOverlayHandler();
-    },
-    [setAppMenuVisible, setSecondaryWithKey, showOverlayHandler]
-  );
-
-  useEffect(() => {
-    if (shouldUpdateFeatureInfo) updateFeatureInfoLeaflet();
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [shouldUpdateFeatureInfo]);
-
-  const handleOnClickTopicMap = useCallback(
-    (e: L.LeafletMouseEvent) => {
-      const map = getTopicMap();
-      if (!map) return;
-
-      if (uiMode === UIMode.FEATURE_INFO) {
-        if (marker !== undefined) {
-          map.removeLayer(marker);
-        }
-        if (markerAccent !== undefined) {
-          map.removeLayer(markerAccent);
-        }
-
-        map.getPane("markerPaneWithBlendModeDifference").style.zIndex = 601;
-        setMarkerAccent(
-          L.marker([e.latlng.lat, e.latlng.lng], {
-            // pane: "backgroundlayerTooltips",
-            icon: L.divIcon({
-              className: "custom-marker", // Optional class for external styles
-              html: `
-                    <div style="
-                      position: relative;
-                      width: 30px;
-                      height: 30px;
-                      border-radius: 50%;
-                      display: flex;
-                      align-items: center;
-                      justify-content: center;
-                      opacity: 1;
-                    ">
-                      <div style="
-                        position: absolute;
-                        width: 20px;
-                        height: 20px;
-                        border: 2px solid black;
-                        border-radius: 50%;
-                      "></div>
-                      <div style="
-                        position: absolute;
-                        width: 20000px;
-                        height: 1px;
-                        background-color: black;
-                        right: 18px;
-                        opacity: 0.5;
-                      "></div>
-                      <div style="
-                        position: absolute;
-                        width: 20000px;
-                        height: 1px;
-                        background-color: black;
-                        left: 18px;
-                        opacity: 0.5;
-                      "></div>
-                      <div style="
-                        position: absolute;
-                        width: 1px;
-                        height: 20000px;
-                        background-color: black;
-                        top: 18px;
-                        opacity: 0.5;
-                      "></div>
-                      <div style="
-                        position: absolute;
-                        width: 1px;
-                        height: 20000px;
-                        background-color: black;
-                        bottom: 18px;
-                        opacity: 0.5;
-                      "></div>
-                    </div>
-                  `,
-              iconSize: [30, 30],
-            }),
-          }).addTo(map)
-        );
-        setMarker(
-          L.marker([e.latlng.lat, e.latlng.lng], {
-            pane: "markerPaneWithBlendModeDifference",
-            icon: L.divIcon({
-              className: "custom-marker", // Optional class for external styles
-              html: `
-                    <div style="
-                      position: relative;
-                      width: 30px;
-                      height: 30px;
-                      border-radius: 50%;
-                      display: flex;
-                      align-items: center;
-                      justify-content: center;
-                    ">
-                       <div style="
-                        position: absolute;
-                        width: 6px;
-                        height: 6px;
-                        background-color: yellow;
-                        border-radius: 50%;
-                      ">
-                    </div>
-                  `,
-              iconSize: [30, 30],
-            }),
-          }).addTo(map)
-        );
-
-        setPos([e.latlng.lat, e.latlng.lng]);
-      }
-      onClickTopicMap(e, {
-        dispatch,
-        mode: uiMode,
-        store,
-        zoom: getLeafletZoom(),
-        map: map,
-      });
-    },
-    [uiMode, marker, markerAccent, dispatch, getLeafletZoom, getTopicMap]
-  );
+  // Feature info update scheduling handled in the wrapper
 
   // TODO Move out Controls to own component
 
@@ -536,153 +128,17 @@ export const GeoportalMap = ({ height, width, allow3d }: MapProps) => {
   lastRenderIntervalRef.current = Date.now() - lastRenderTimeStampRef.current;
   lastRenderTimeStampRef.current = Date.now();
 
-  const modalMenu = useMemo(
-    () => (
-      <GenericModalApplicationMenu
-        {...getCollabedHelpComponentConfig({
-          versionString: version,
-          showOverlayFromOutside,
-          loginFormToggle: () => setIsLoginFormVisible(!isLoginFormVisible),
-          isLoginFormVisible,
-          loginForm: (
-            <LoginForm
-              onSuccess={() => {
-                setIsLoginFormVisible(false);
-                setAppMenuVisible(false);
-              }}
-              closeLoginForm={() => setIsLoginFormVisible(false)}
-            />
-          ),
-          loginFormTrigger: (
-            <Tooltip title={jwt ? "Abmeldung" : "Anmeldung"} zIndex={99999999}>
-              <Button
-                type="text"
-                onClick={() =>
-                  jwt
-                    ? setJWT(null)
-                    : setIsLoginFormVisible(!isLoginFormVisible)
-                }
-              >
-                <FontAwesomeIcon
-                  icon={jwt ? faArrowRightFromBracket : faKey}
-                  size="lg"
-                />
-              </Button>
-            </Tooltip>
-          ),
-        })}
-      />
-    ),
-    [
-      version,
-      showOverlayFromOutside,
-      isLoginFormVisible,
-      jwt,
-      setJWT,
-      setAppMenuVisible,
-    ]
-  );
-
-  const topicMapLayers = useCreateCismapLayers(layers, {
-    mode: uiMode,
-    dispatch,
-    zoom: getLeafletZoom(),
-    selectedFeature,
-    leafletMap: getTopicMap(),
-  });
-
-  const topicMapBackgroundLayer = useMemo(() => {
-    return (
-      backgroundLayer &&
-      backgroundLayer.visible &&
-      getBackgroundLayers({
-        layerString: backgroundLayer.layers,
-        masterOpacity: backgroundLayer.opacity,
-      })
-    );
-  }, [backgroundLayer]);
-
-  const memoizedTopicMap = useMemo(() => {
-    console.info("RENDER: [GEOPORTAL] MEMOIZED TOPICMAP");
-    return (
-      <TopicMapComponent
-        gazData={gazData}
-        modalMenu={modalMenu}
-        gazetteerSearchComponent={EmptySearchComponent}
-        applicationMenuTooltipString={tooltipText}
-        hamburgerMenu={showHamburgerMenu}
-        locatorControl={false}
-        fullScreenControl={false}
-        zoomControls={false}
-        mapStyle={{
-          width,
-          height,
-          touchAction: "none",
-          WebkitOverflowScrolling: "touch",
-          overscrollBehavior: "none",
-        }}
-        leafletMapProps={{ editable: true }}
-        minZoom={10}
-        backgroundlayers="empty"
-        mappingBoundsChanged={() => {
-          // intentionally no-op
-        }}
-        locationChangedHandler={topicMapLocationChangedHandler}
-        outerLocationChangedHandlerExclusive={true}
-        onclick={handleOnClickTopicMap}
-        gazetteerSearchControl={true}
-        infoBox={renderInfoBox()}
-        zoomSnap={LEAFLET_CONFIG.zoomSnap}
-        zoomDelta={LEAFLET_CONFIG.zoomDelta}
-      >
-        <TopicMapSelectionContent />
-        {topicMapBackgroundLayer}
-        {topicMapLayers}
-        <PrintPreview />
-      </TopicMapComponent>
-    );
-  }, [
-    handleOnClickTopicMap,
-    topicMapBackgroundLayer,
-    topicMapLayers,
-    gazData,
-    modalMenu,
-    showHamburgerMenu,
-    topicMapLocationChangedHandler,
-    renderInfoBox,
-    width,
-    height,
-  ]);
-
   return (
     <>
-      <div className={"map-container-2d"} style={{ zIndex: 400 }}>
-        {memoizedTopicMap}
-      </div>
-      {allow3d && cesiumInitialCameraView !== null && (
-        <div
-          ref={container3dMapRef}
-          className={"map-container-3d"}
-          style={{
-            position: "absolute",
-            top: 0,
-            left: 0,
-            right: 0,
-            bottom: 0,
-            zIndex: 400,
-            opacity: isMode2d ? 0 : 1,
-            transition: `opacity ${CESIUM_CONFIG.transitions.mapMode.duration}ms ease-in-out`,
-            pointerEvents: isMode2d ? "none" : "auto",
-          }}
-        >
-          <CustomViewer
-            containerRef={container3dMapRef}
-            cameraLimiterOptions={CESIUM_CONFIG.camera}
-            initialCameraView={cesiumInitialCameraView}
-            onSceneChange={cesiumLocationChangedHandler}
-          />
-        </div>
-      )}
+      <TopicMapComponentWrapper
+        locationChangedHandler={topicMapLocationChangedHandler}
+        leafletOptions={leafletOptions}
+      />
+      <CesiumMapComponentWrapper
+        allow3d={allow3d}
+        cesiumOptions={cesiumOptions}
+        onSceneChange={cesiumLocationChangedHandler}
+      />
     </>
   );
 };
