+++ apps/geoportal/src/app/components/GeoportalMap/components/TopicMapComponentWrapper/index.ts
@@ -1,0 +1,1 @@
+export { TopicMapComponentWrapper } from "./TopicMapComponentWrapper";
