+++ apps/geoportal/src/app/components/GeoportalMap/utils/useOverlayHelpers.ts
@@ -1,0 +1,29 @@
+import { useOverlayHelper } from "@carma-commons/ui/helper-overlay";
+import { geoElements } from "@carma-collab/wuppertal/geoportal";
+import { getCollabedHelpComponentConfig as getCollabedHelpElementsConfig } from "@carma-collab/wuppertal/helper-overlay";
+
+import { addCssToOverlayHelperItem } from "../../../helper/overlayHelper.ts";
+
+export const useGeoportalOverlays = () => {
+  const infoBoxOverlay = addCssToOverlayHelperItem(
+    getCollabedHelpElementsConfig("INFOBOX", geoElements),
+    "350px",
+    "137px"
+  );
+
+  const layerButtonsOverlay = addCssToOverlayHelperItem(
+    getCollabedHelpElementsConfig("LAYERBUTTONS", geoElements),
+    "146px",
+    "21px"
+  );
+
+  const mapInteractionOverlay = addCssToOverlayHelperItem(
+    getCollabedHelpElementsConfig("CENTER", geoElements),
+    "15px",
+    "15px"
+  );
+
+  useOverlayHelper(infoBoxOverlay);
+  useOverlayHelper(layerButtonsOverlay);
+  useOverlayHelper(mapInteractionOverlay);
+};
